frontend/pyproject.toml
[project]
name = "p-url-frontend"
version = "0.1.0"
description = "Streamlit frontend for url shortener for code challenge"
readme = "README.md"
requires-python = ">=3.11"
dependencies = ["streamlit", "python-dotenv", "streamlit_js_eval"]

[tool.uv]
dev-dependencies = [
    "pytest",
    "pytest-testdox",
    "pytest-cov",
    "coverage",
    "bandit",
    "pip-audit",
    "black",
    "flake8",
    "flake8-pyproject",
]

[tool.pytest.ini_options]
addopts = "--testdox"
testpaths = ["tests"]
pythonpath = "./"

[tool.flake8]
max-line-length = 100
exclude = [
    ".git",
    "__pycache__",
    ".venv",
    "venv",
    "env",
    "build",
    "dist"
]

[tool.black]
line-length = 100

[tool.bandit]
exclude_dirs = ["./tests", "./.venv", "./build"]
skips = ["B101"]
